requirements.txt
python-telegram-bot
python-dotenv
azure-cognitiveservices-vision-computervision
Pillow